Refactor feature toggles to be more Rails like
It seems as though application logic should not live in the config directory so FeatureFlags has moved to lib.
